Auto-focus text input when user presses A-z
It's useful to auto focus the input file when the user presses a key,
so they don't have to move the mouse after the page is loaded. We only
do this 500 ms after page is loaded, so it doesn't get pressed
accidentally (e.g. if the user reloaded the page with CMD-R and the page
is still processing a reload event)